backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo==4.6.1
pinecone-client==2.2.1
langchain==0.0.267
openai==0.28.1
tiktoken==0.5.2
pypdf==3.15.1 